is-lnu-service/src/test/java/org/lnu/is/extractor/street/type/StreetTypeParametersExtractorTest.java
package org.lnu.is.extractor.street.type;

import static org.junit.Assert.assertEquals;

import java.util.HashMap;
import java.util.Map;

import org.junit.Test;
import org.lnu.is.domain.streettype.StreetType;


public class StreetTypeParametersExtractorTest {

	private StreetTypeParametersExtractor unit = new StreetTypeParametersExtractor();
	
	@Test
	public void testGetParameters() throws Exception {
		// Given
		String name = "Extract Me All";
		String abbrName = "EML";

		StreetType entity = new StreetType();
		entity.setName(name);
		entity.setAbbrName(abbrName);

		Map<String, Object> expected = new HashMap<String, Object>();
		expected.put("name", name);
		expected.put("abbrName", abbrName);

		// When
		Map<String, Object> actual = unit.getParameters(entity);

		// Then
		assertEquals(expected, actual);
	}
}
